Move actual form stuff to ReviewForm
