+++ src/modules/CourseDetail/context/Review/constants.ts
@@ -3,6 +3,7 @@ import { createRef } from 'react'
 import { ReviewContextValues } from './types'
 
 export const DEFAULT_REVIEW_CONTEXT_VALUE: ReviewContextValues = {
+  courseNo: '',
   reviews: [],
   myPendingReviews: [],
   setInteraction: () => {},
@@ -12,7 +13,9 @@ export const DEFAULT_REVIEW_CONTEXT_VALUE: ReviewContextValues = {
   cancelEditReview: () => {},
   submitReview: () => {},
   submitEditedReview: () => {},
-  editorRef: createRef(),
+  formRef: createRef(),
+  formLoaded: false,
+  onFormLoad: () => {},
 }
 
 export const REVIEW_FORM_ID = 'REVIEW_FORM'
